Support a gateway user without SSH keys
Add support for an Aker user who does not have SSH keys, perhaps because
password-based authentication is used.  Accomplished by catching an
invalid SSH key and assuming that password fallback is sane.

Adds a check later to determine if the secret authentication mechanism
is a string or not.  If it is a string, it's assumed that password
authentication is desired.