Handle potential EOFError with network devices.
Some network device operating systems (looking at you, IOS!) may cause
an EOFError to be raised.  This change catches that possibility and
assumes that it is a normal occurence, logging it and cleaning up the
connection.

Connection cleanup logic also gets its own function since it is and will
be used as the number of exceptions possible increases.